Extracted target dir name generation
