+++ conf/script/src/build_system/cmd/hierarchy/create_target_build_dirs/target_dir_name_generation.py
