+++ conf/script/src/build_system/cmd/setup/impl.py
@@ -9,7 +9,7 @@ def setup_build_system(root_dir: Optional[Path] = None):
     import build_system.cmd.hierarchy.create_build_dir
     import build_system.cmd.hierarchy.find_root_dir
     import build_system.cmd.hierarchy.find_build_dir
-    import build_system.cmd.setup.build_name
+    import build_system.cmd.hierarchy.create_target_build_dirs.target_dir_name_generation
     import build_system.cmd.setup.create_build_subdirs
 
     if root_dir is None:
@@ -22,5 +22,5 @@ def setup_build_system(root_dir: Optional[Path] = None):
     build_system.cmd.hierarchy.clean_build_dir.clean_build_dir(build_dir=build_dir, ignore_errors=True)
     build_dir = build_system.cmd.hierarchy.create_build_dir.create_build_dir(root_dir=root_dir)
 
-    all_build_subdir_names = build_system.cmd.setup.build_name.generate_all_build_subdir_names()
+    all_build_subdir_names = build_system.cmd.hierarchy.create_target_build_dirs.target_dir_name_generation.generate_all_build_subdir_names()
     build_system.cmd.setup.create_build_subdirs.create_all_build_subdirs(build_dir, all_build_subdir_names)
